+++ src/perspectiveCamera.cpp
@@ -7,7 +7,7 @@ PerspectiveCamera::PerspectiveCamera(){}
 
 PerspectiveCamera::PerspectiveCamera(
     Point e, Vector3 u, Vector3 v, Vector3 w,
-    std::tuple<float, float, float, float> screenWindow
+    std::tuple<double, double, double, double> screenWindow
 ) : Camera(e, u, v, w, screenWindow) {}
 
 Ray PerspectiveCamera::generate_ray(int x, int y)
