+++ tests/test_ray.cpp
@@ -15,7 +15,7 @@ TEST_CASE("it can positionate ray") {
     Point origin(0, 0, 0);
     Vector3 direction(3, 2, -4);
     Ray r(origin, direction);
-    float t = 2.3;
+    double t = 2.3;
     Point result(6.8999999999999995, 4.6, -9.2);
     CHECK_EQ(r(t), result);
 }
